Remove PushRegistrationIntentService and replace with Runnable tasks
If pushRegister is called too soon in the application lifecycle, the process
may still be considered in a background state and then the background service
restrictions still apply.

By refactoring to runnables handed to the executor, these should function so
long as the process is alive (regardless of state).
